rename fetchInitiative --> getInitiative
